+++ options/defaults.lua
@@ -487,36 +487,44 @@ SAO.defaults = {
                 [17941] = { -- Nightfall
                     [0] = true,
                 },
-                [34936] = { -- Backlash
+                [34936] = { -- Backlash (TBC+)
                     [0] = true,
                 },
-                [71165] = { -- Molten Core
+                [71165] = { -- Molten Core (Wrath+)
                     [0] = true, -- any stacks
                 },
-                [63167] = { -- Decimation
+                [63167] = { -- Decimation (Wrath+)
                     [0] = true,
                 },
-                [47283] = { -- Empowered Imp
+                [47283] = { -- Empowered Imp (Wrath+)
                     [0] = true,
                 },
             },
             glow = {
+                [1120] = { -- Drain Soul (Wrath+)
+                    [1120] = "spec:1", -- Drain Soul
+                },
                 [17941] = { -- Nightfall
                     [686] = true, -- Shadow Bolt
                 },
-                [34936] = { -- Backlash
+                [71165] = { -- Molten Core (Wrath+)
+                    [29722] = true, -- Incinerate
+                    [6353]  = true, -- Soul Fire (not for Cata)
+                },
+                [63167] = { -- Decimation (Wrath+)
+                    [6353] = true, -- Soul Fire
+                },
+                [54277] = { -- Backdraft (Cataclysm)
+                    [686]   = false, -- Shadow Bolt
+                    [29722] = false, -- Incinerate
+                    [50796] = false, -- Chaos Bolt
+                },
+                [34936] = { -- Backlash (TBC+)
                     [686]   = true, -- Shadow Bolt
                     [29722] = true, -- Incinerate
                 },
-                [71165] = { -- Molten Core
-                    [29722] = true, -- Incinerate
+                [47283] = { -- Empowered Imp (Cataclysm)
                     [6353]  = true, -- Soul Fire
-                },
-                [63167] = { -- Decimation
-                    [6353] = true, -- Soul Fire
-                },
-                [1120] = { -- Drain Soul (Wrath+)
-                    [1120] = "spec:1", -- Drain Soul
                 },
             },
         },